LPS-130175 Move ResizeContext component
