+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/plugins/page-structure/components/item-configuration-panels/RowStylesPanel.test.js
@@ -16,8 +16,8 @@ import '@testing-library/jest-dom/extend-expect';
 import {cleanup, fireEvent, render} from '@testing-library/react';
 import React from 'react';
 
-import {ResizeContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/components/ResizeContext';
 import {config} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/config/index';
+import {ResizeContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/contexts/ResizeContext';
 import {StoreAPIContextProvider} from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/store/index';
 import updateItemConfig from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateItemConfig';
 import updateRowColumns from '../../../../../../../src/main/resources/META-INF/resources/page_editor/app/thunks/updateRowColumns';
